move KMS initialization to single place
