+++ src/encrypt-api/Startup.cs
@@ -55,52 +55,11 @@ namespace Kamus
                 swagger.SwaggerDoc ("v1", new Swashbuckle.AspNetCore.Swagger.Info { Title = "Kamus Swagger" });
             });
 
-            services.AddScoped<IKeyManagement>(s =>
-            {
-                var provider = Configuration.GetValue<string>("KeyManagement:Provider");
-                switch (provider)
-                {
-                    case "AwsKms":
-                        return GetAwsKeyManagement(Log.Logger);
-                    case "GoogleKms":
-                        return GetGoogleCloudKeyManagment();
-                    case "AzureKeyVault":
-                        return new EnvelopeEncryptionDecorator(
-                            new AzureKeyVaultKeyManagement(s.GetService<IKeyVaultClient>(), Configuration), 
-                            new SymmetricKeyManagement(),
-                            Configuration.GetValue<int>("KeyManagement:KeyVault:MaximumDataLength"));
-                    case "AESKey":
-                        var key = Configuration.GetValue<string>("KeyManagement:AES:Key");
-                        if (string.IsNullOrEmpty(key))
-                        {
-                            Log.ForContext<Startup>().Warning("Random key was created for SymmetricKeyManagement, it might break distributed deployments");
-                        }
-                        return new SymmetricKeyManagement(key);
-                    default:
-                        throw new InvalidOperationException($"Unsupported provider type: {provider}");
-                }
-            });
-
-            services.AddSingleton<IKeyVaultClient>(_ => new KeyVaultClient(GetToken));
+            services.AddKeyManagement(Configuration);
 
             services.AddSingleton<IConfiguration>(Configuration);
 
             services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-        }
-
-        public async Task<string> GetToken(string authority, string resource, string scope)
-        {
-            var clientId = Configuration["ActiveDirectory:ClientId"];
-            var clientSecret = Configuration["ActiveDirectory:ClientSecret"];
-            
-            var authContext = new AuthenticationContext(authority);
-            ClientCredential clientCred = new ClientCredential(clientId, clientSecret);
-            AuthenticationResult result = await authContext.AcquireTokenAsync(resource, clientCred);
-
-            if (result == null)
-                throw new InvalidOperationException("Failed to obtain the JWT token");
-
-            return result.AccessToken;
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -128,59 +87,6 @@ namespace Kamus
             app.UseAuthentication();
 
             app.UseMvc ();
-        }
-
-        private IKeyManagement GetGoogleCloudKeyManagment()
-        {
-            var location = Configuration.GetValue<string>("KeyManagement:GoogleKms:Location");
-            var keyRingName = Configuration.GetValue<string>("KeyManagement:GoogleKms:KeyRingName");
-            var protectionLevel = Configuration.GetValue<string>("KeyManagement:GoogleKms:ProtectionLevel");
-            var credentialsPath = Configuration.GetValue<string>("KeyManagement:GoogleKms:CredentialsPath");
-
-            var serviceAccountCredential = ServiceAccountCredential.FromServiceAccountData(File.OpenRead(credentialsPath));
-            var credentials = GoogleCredential.FromServiceAccountCredential(serviceAccountCredential);
-            if (credentials.IsCreateScopedRequired)
-            {
-                credentials = credentials.CreateScoped(new[]
-                {
-                    CloudKMSService.Scope.CloudPlatform
-                });
-            }
-
-            var kmsService = new CloudKMSService(new BaseClientService.Initializer
-            {
-                HttpClientInitializer = credentials,
-                GZipEnabled = true
-            });
-
-
-            return new GoogleCloudKeyManagment(
-                kmsService,
-                serviceAccountCredential.ProjectId,
-                keyRingName,
-                location,
-                protectionLevel);
-        }
-        
-        private IKeyManagement GetAwsKeyManagement(ILogger logger)
-        {
-            AmazonKeyManagementServiceClient kmsService;
-            var region = Configuration.GetValue<string>("KeyManagement:AwsKms:Region");
-            var awsKey = Configuration.GetValue<string>("KeyManagement:AwsKms:Key");
-            var awsSecret = Configuration.GetValue<string>("KeyManagement:AwsKms:Secret");
-            var cmkPrefix = Configuration.GetValue<string>("KeyManagement:AwsKms:CmkPrefix");
-            
-            if (string.IsNullOrEmpty(region) || string.IsNullOrEmpty(awsKey) || string.IsNullOrEmpty(awsSecret))
-            {
-                logger.Information("AwsKms credentials were not provided, using default AWS SDK credentials discovery");
-                kmsService = new AmazonKeyManagementServiceClient();
-            }
-            else
-            {
-                kmsService = new AmazonKeyManagementServiceClient(awsKey, awsSecret, RegionEndpoint.GetBySystemName(region));
-            }
-            
-            return new AwsKeyManagement(kmsService, new SymmetricKeyManagement(), cmkPrefix);
         }
     }
 }
